+++ apps/cli/src/storage.ts
@@ -11,9 +11,20 @@ export async function setSessionSecret(sessionSecret: string): Promise<void> {
   await userSettingsJsonFile().setAsync('sessionSecret', sessionSecret);
 }
 
+export function getTrustedSources(): string[] | undefined {
+  const userSettings = userSettingsJsonFile().read();
+  return userSettings.trustedSources;
+}
+
+export async function setTrustedSources(trustedSources: string[] | undefined): Promise<void> {
+  await userSettingsJsonFile().setAsync('trustedSources', trustedSources);
+}
+
 type UserData = {
   sessionSecret?: string;
+  trustedSources?: string[];
 };
+
 function userSettingsJsonFile(): JsonFile<UserData> {
   return new JsonFile<UserData>(StorageUtils.userSettingsFile(os.homedir()), {
     jsonParseErrorDefault: {},