+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### 🎉 New features
 
+- Add support for configuring trusted sources for opening URLs via the CLI and menu-bar app. ([#270](https://github.com/expo/orbit/pull/270) by [@altaywtf](https://github.com/altaywtf))
+
 ### 🐛 Bug fixes
 
 ### 💡 Others